chore (core) move pattern related classes to their own folder
