+++ packages/core/src/index.mjs
@@ -1,13 +1,13 @@
 import { Bezier } from 'bezier-js'
 import { Attributes } from './attributes.mjs'
 import { Design } from './design.mjs'
-import { Pattern } from './pattern.mjs'
+import { Pattern } from './pattern/index.mjs'
 import { Part } from './part.mjs'
 import { Point } from './point.mjs'
 import { Path } from './path.mjs'
 import { Snippet } from './snippet.mjs'
 import { Store } from './store.mjs'
-import { hidePresets } from './pattern-config.mjs'
+import { hidePresets } from './pattern/pattern-config.mjs'
 import {
   beamIntersectsCircle,
   beamIntersectsX,
