+++ packages/core/src/pattern/index.mjs
@@ -1,16 +1,16 @@
-import { Attributes } from './attributes.mjs'
+import { Attributes } from '../attributes.mjs'
 import pack from 'bin-pack-with-constraints'
-import { __addNonEnumProp, __macroName } from './utils.mjs'
-import { Part } from './part.mjs'
-import { Stack } from './stack.mjs'
-import { Point } from './point.mjs'
-import { Path } from './path.mjs'
-import { Snippet } from './snippet.mjs'
-import { Svg } from './svg.mjs'
-import { Store } from './store.mjs'
-import { Hooks } from './hooks.mjs'
-import { version } from '../data.mjs'
-import { __loadPatternDefaults } from './config.mjs'
+import { __addNonEnumProp, __macroName } from '../utils.mjs'
+import { Part } from '../part.mjs'
+import { Stack } from '../stack.mjs'
+import { Point } from '../point.mjs'
+import { Path } from '../path.mjs'
+import { Snippet } from '../snippet.mjs'
+import { Svg } from '../svg.mjs'
+import { Store } from '../store.mjs'
+import { Hooks } from '../hooks.mjs'
+import { version } from '../../data.mjs'
+import { __loadPatternDefaults } from '../config.mjs'
 import { PatternConfig, getPluginName } from './pattern-config.mjs'
 import { PatternDraftQueue } from './pattern-draft-queue.mjs'
 import cloneDeep from 'lodash.clonedeep'
